Add cached image network
