+++ lib/features/parts_with_material_number/screens/all_parts_screen.dart
@@ -1,3 +1,4 @@
+import 'package:cached_network_image/cached_network_image.dart';
 import 'package:conditional_builder_null_safety/conditional_builder_null_safety.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -12,6 +13,7 @@ import 'package:rollshop/core/theme/styles.dart';
 import 'package:rollshop/features/parts_with_material_number/model/parts_with_material_number_model.dart';
 import 'package:rollshop/features/parts_with_material_number/view_model/cubit/parts_cubit.dart';
 import 'package:rollshop/features/parts_with_material_number/view_model/cubit/parts_state.dart';
+import 'package:rollshop/features/parts_with_material_number/widgets/build_part_item.dart';
 import 'package:rollshop/injection_container.dart';
 
 class AllPartsScreen extends StatefulWidget {
@@ -111,91 +113,4 @@ class _AllPartsScreenState extends State<AllPartsScreen> {
       },
     );
   }
-}
-
-class BuildPartItem extends StatelessWidget {
-  const BuildPartItem({
-    super.key,
-    required this.part,
-    required this.allowEdit,
-  });
-
-  final PartsWithMaterialNumberModel part;
-  final bool allowEdit;
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: () {
-        if (!allowEdit) return;
-        debugPrint("Pressed");
-        context.pushNamed(Routes.editPartScreen, arguments: part);
-      },
-      child: Padding(
-        padding: EdgeInsets.all(10.sp),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          spacing: 20,
-          children: [
-            Expanded(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  buildPartItem(text: part.name, lable: "الأسم:"),
-                  buildPartItem(
-                      text: part.materialNumber, lable: "رقم ماتريال:"),
-                  buildPartItem(
-                      text: part.drawingPartNumber, lable: "رقم الرسمة"),
-                  buildPartItem(text: part.sizes, lable: "المقاسات"),
-                  buildPartItem(text: part.type, lable: "النوع"),
-                ],
-              ),
-            ),
-            SizedBox(
-              width: 120.w,
-              height: 120.h,
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(15),
-                child: part.imagePath.isNotEmpty
-                    ? BuildImageWithErrorHandler(
-                        imageType: ImageType.network,
-                        // path: parts[index].imagePath,
-                        path: part.imagePath,
-                        boxFit: BoxFit.cover,
-                      )
-                    : SizedBox(
-                        child: Center(
-                          child: Text(
-                            "No Image",
-                          ),
-                        ),
-                      ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-Widget buildPartItem({required dynamic text, required String lable}) {
-  return Row(
-    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-    children: [
-      Text(
-        lable,
-        style: MyTextStyles.font12BlackeBold,
-      ),
-      Flexible(
-        child: Text(
-          text.toString(),
-          textDirection: TextDirection.ltr,
-          overflow: TextOverflow.ellipsis,
-          maxLines: 2,
-        ),
-      ),
-    ],
-  );
 }
